tx/types: rename tx_snapshot_utils to rm_stm_types
Next change consolidates more types in this header
