+++ src/v/cluster/rm_stm.cc
@@ -12,8 +12,8 @@
 #include "bytes/iostream.h"
 #include "cluster/logger.h"
 #include "cluster/producer_state_manager.h"
+#include "cluster/rm_stm_types.h"
 #include "cluster/tx_gateway_frontend.h"
-#include "cluster/tx_snapshot_utils.h"
 #include "cluster/types.h"
 #include "kafka/protocol/wire.h"
 #include "metrics/metrics.h"
