+++ src/v/cluster/rm_stm_types.cc
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "cluster/tx_snapshot_utils.h"
+#include "cluster/rm_stm_types.h"
 
 namespace cluster {
