+++ src/v/cluster/rm_stm_types.h
