+++ src/v/cluster/tests/rm_stm_tests.cc
@@ -8,9 +8,9 @@
 // by the Apache License, Version 2.0
 
 #include "cluster/errc.h"
+#include "cluster/rm_stm_types.h"
 #include "cluster/tests/randoms.h"
 #include "cluster/tests/rm_stm_test_fixture.h"
-#include "cluster/tx_snapshot_utils.h"
 #include "finjector/hbadger.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
